pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "safeguard-toolkit"
version = "0.1.5"
description = "A toolkit for scanning configs, secrets, and permissions."
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Purvi Verma", email = "20ce01050@iitbbs.ac.in"}
]
dependencies = [
        "pyyaml==6.0.2",
        "requests==2.32.3",
        "packaging==25.0",
        "toml==0.10.2",
        "packaging==25.0"
    ]

[project.urls]
Homepage = "https://github.com/purvi1508/Safeguard"
